@@ -2,15 +2,15 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `rails
+# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_02_21_221723) do
+ActiveRecord::Schema.define(version: 2020_03_17_164157) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -103,14 +103,12 @@ ActiveRecord::Schema.define(version: 2020_02_21_221723) do
     t.datetime "stop"
     t.text "description"
     t.text "terms_of_service"
-    t.boolean "disable_flags_an_hour_graph", default: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "organization"
     t.string "contact_url"
     t.text "footer"
     t.integer "team_size", default: 5
-    t.string "do_not_reply_email"
     t.string "contact_email"
     t.string "open_source_url"
     t.boolean "prizes_available", default: false
